Implemented GetLAtestGeneration and Get for changelog + one test.

+++ src/DotJEM.Json.Storage2/SqlServer/Statements/CreateLogTable.sql
@@ -4,7 +4,7 @@
     [Event] [varchar](1) NOT NULL,
     [Time] [datetime2] NOT NULL,
     [User] [nvarchar](256) NULL, 
-    [Version] [bigint] NOT NULL,
+    [Version] [int] NOT NULL,
     [Data] [nvarchar](max) NOT NULL,
     CONSTRAINT [PK_@{area_name}.log] PRIMARY KEY CLUSTERED ( [Revision] ASC )
     WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
